% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genevaR.R
\name{genevaR}
\alias{genevaR}
\title{Gene and Variants Plot}
\usage{
genevaR(gff_file, gene_name, vcf_file = NULL)
}
\arguments{
\item{gff_file}{Data frame including columns "seqid", "start", "end", "type" and "ID".}

\item{gene_name}{ID of the gene of interest. Should be the same for the GFF and VCF input.}

\item{vcf_file}{.txt file including columns "POS", "ANN....GENEID" (gene ID) and "ANN....EFFECT" (variant type). SnpSift-filtered VCF file.}
}
\description{
Creates a plot that visualizes the selected gene and optionally marks variant positions.
}
\keyword{exons}
\keyword{gene}
\keyword{gff}
\keyword{plot}
\keyword{snpsift}
\keyword{variants}
\keyword{vcf}
